pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "agenttrust"
version = "0.1.0"
description = "AgentTrust Protocol - Secure payment infrastructure for AI agents"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pynacl>=1.5.0",
    "pyjwt[crypto]>=2.8.0",
    "cryptography>=41.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.19.0",
    "stripe>=7.0.0",
    "httpx>=0.26.0",
    "pydantic>=2.0.0",
    "base58>=2.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
sdk = [
    "langchain>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["server*", "sdk*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
